tests/test_main.py
import unittest
import logging
import os
from unittest.mock import patch, MagicMock
import pandas as pd
from scripts.main import execute_sql_script, load_csv_to_db, process_dataset, main, setup_logging

class TestMainFunctions(unittest.TestCase):

    def setUp(self):
        # Create a temporary test database or use a mock database for testing
        self.test_db_name = 'test_db'
        self.test_conn = MagicMock()
        
    def tearDown(self):
        # Clean up any resources after each test if needed
        pass

    def test_execute_sql_script(self):
        # Mocking the SQL script execution
        sql_file = 'scripts/sql_integration/init.sql'
        
        with patch('psycopg2.connect') as mock_connect:
            mock_conn = mock_connect.return_value
            mock_cursor = mock_conn.cursor.return_value
            
            execute_sql_script(mock_conn, sql_file)
            
            mock_cursor.execute.assert_called_once()
            mock_conn.commit.assert_called_once()

    def test_load_csv_to_db(self):
        # Mocking the CSV loading into database
        csv_file = 'test.csv'
        table_name = 'test_table'
        test_data = pd.DataFrame({'col1': [1, 2], 'col2': ['a', 'b']})

        with patch('psycopg2.connect') as mock_connect:
            mock_conn = mock_connect.return_value
            mock_cursor = mock_conn.cursor.return_value

            load_csv_to_db(mock_conn, csv_file, table_name)
            
            mock_cursor.executemany.assert_called_once()
            mock_conn.commit.assert_called_once()

    def test_process_dataset(self):
        # Test the dataset processing function
        df1 = pd.DataFrame({'order_id': [1, 2, 3], 'lat': [0, 1, 2], 'lng': [0, 1, 2]})
        df2 = pd.DataFrame({'Trip ID': [1, 2, 3], 'Trip Start Time': ['2023-01-01 08:00:00', '2023-01-02 12:00:00', '2023-01-03 18:00:00'], 'Trip End Time': ['2023-01-01 09:00:00', '2023-01-02 13:00:00', '2023-01-03 19:00:00'], 'Trip Origin': ['0.0,0.0', '1.0,1.0', '2.0,2.0'], 'Trip Destination': ['1.0,1.0', '2.0,2.0', '3.0,3.0']})
        
        try:
            processed_df = process_dataset(df1, df2)
            self.assertIsInstance(processed_df, pd.DataFrame)
        except Exception as e:
            self.fail(f"process_dataset raised {type(e)}: {str(e)}")

    @patch('main.setup_logging')
    def test_main(self, mock_setup_logging):
        # Test the main function (mocking setup_logging)
        mock_setup_logging.return_value = None
        
        try:
            main()
            # Add assertions as per your requirements (e.g., check printed outputs)
            # Mock other function calls within main if needed
        except Exception as e:
            self.fail(f"main raised {type(e)}: {str(e)}")

    def test_setup_logging(self):
        # Test the setup_logging function
        try:
            setup_logging()
            # Check if logging configuration is set correctly
            self.assertEqual(logging.getLogger().level, logging.INFO)
        except Exception as e:
            self.fail(f"setup_logging raised {type(e)}: {str(e)}")

if __name__ == '__main__':
    unittest.main()

